+++ src/preload/index.d.ts
@@ -62,7 +62,7 @@ export interface ApplicationAPI {
   isProjectPath: (path: string) => Promise<boolean>;
   dropFile: (baseDir: string, path: string) => void;
   runCommand: (baseDir: string, command: string) => void;
-  scrapeWeb: (baseDir: string, url: string) => Promise<string>;
+  scrapeWeb: (baseDir: string, url: string, filePath?: string) => Promise<string>;
   initProjectRulesFile: (baseDir: string) => Promise<void>;
 
   // Todo operations