+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added custom commands availability to all modes
 - added info message when agent hits max output tokens
 - improved handling of unknown finsh reason during agent iteration
+- fixed UnicodeEncodeError on Windows inside Aider
 
 ## [0.21.0]