+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added info message when agent hits max output tokens
 - improved handling of unknown finsh reason during agent iteration
 - fixed UnicodeEncodeError on Windows inside Aider
+- correctly handling stored answer
 
 ## [0.21.0]